This is a jajor re-write of the programming algorithms.  The Atmel serial programming instructions are not very orthoganal, i.e., the "read fuse bits" instruction on an ATMega103 is an entirely different opcode and data format from the _same_ instruction for an ATMega163! Thus, it becomes impossible to have a single instruction encoding (varying the data) across the chip lines.
This set of changes allows and requires instruction encodings to be
defined on a per-part basis within the configuration file.  Hopefully
I've defined the encoding scheme in a general enough way so it is
useful in describing the instruction formats for yet-to-be invented
Atmel chips.  I've tried hard to make it match very closely with the
specification in Atmel's data sheets for their parts.  It's a little
more verbose than what I initially hoped for, but I've tried to keep
it as concise as I could, while still remaining reasonably flexible.


git-svn-id: svn://svn.savannah.nongnu.org/avrdude/trunk/avrdude@100 81a1dc3b-b13d-400b-aceb-764788c761c2